# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "description": "Stable machine-readable error code.",
            "pattern": "^[A-Z_]+$",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "description": "Safe human-readable explanation; never raw server details.",
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "JobAcceptedResponse": {
        "properties": {
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/JobState"
          },
          "status_url": {
            "title": "Status Url",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "state",
          "status_url"
        ],
        "title": "JobAcceptedResponse",
        "type": "object"
      },
...
      "JobState": {
        "enum": [
          "PENDING",
          "QUEUED",
          "STARTED",
          "COMPLETED",
          "FAILED",
          "CANCELLED"
        ],
        "title": "JobState",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "JWT access token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Domain Copilot HTTP API. Bearer JWT identity and stored roles are checked server-side. See docs/API-CONTRACTS.md for ownership and SSE semantics.",
    "title": "Domain Copilot",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/evaluations": {
      "post": {
        "description": "Admin-only submission, with no request body. The server pins its configured catalog.\n\nReturns a durably accepted job and Location/status_url before evaluation executes.\n\nRoles: admin. Owner only for analyst/reviewer; admin may access all owned jobs. Identity and current role come from the validated JWT and stored user.",
        "operationId": "submit_evaluation_api_v1_evaluations_post",
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAcceptedResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "Location": {
                "description": "Relative resource/status URL.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "DOMAIN_ERROR, APPLICATION_ERROR or HTTP_ERROR: invalid operation."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "NOT_AUTHENTICATED. WWW-Authenticate: Bearer.",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "const": "Bearer",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "PERMISSION_DENIED or RESOURCE_FORBIDDEN."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RESOURCE_NOT_FOUND or HTTP_ERROR."
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "HTTP_ERROR: method not allowed."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INVALID_STATE_TRANSITION: conflict with persisted state."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "UPLOAD_TOO_LARGE."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "VALIDATION_ERROR or INVARIANT_VIOLATION: invalid input."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INTERNAL_ERROR or CONFIGURATION_ERROR. Static, sanitized detail."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "JOB_STORE_UNAVAILABLE, KNOWLEDGE_UNAVAILABLE, APPROVAL_STORE_UNAVAILABLE or HISTORY_STORE_UNAVAILABLE. Static, sanitized detail."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Submit Evaluation",
        "tags": [
          "evaluations"
        ],
        "x-ownership": "Owner only for analyst/reviewer; admin may access all owned jobs.",
        "x-roles": [
          "admin"
        ]
      }
    },
    "/api/v1/evaluations/{job_id}": {
      "get": {
        "description": "Roles: analyst, reviewer, admin. Owner only for analyst/reviewer; admin may access all owned jobs. Identity and current role come from the validated JWT and stored user.",
        "operationId": "evaluation_status_api_v1_evaluations__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Evaluation Status Api V1 Evaluations  Job Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "DOMAIN_ERROR, APPLICATION_ERROR or HTTP_ERROR: invalid operation."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "NOT_AUTHENTICATED. WWW-Authenticate: Bearer.",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "const": "Bearer",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "PERMISSION_DENIED or RESOURCE_FORBIDDEN."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RESOURCE_NOT_FOUND or HTTP_ERROR."
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "HTTP_ERROR: method not allowed."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INVALID_STATE_TRANSITION: conflict with persisted state."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "UPLOAD_TOO_LARGE."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "VALIDATION_ERROR or INVARIANT_VIOLATION: invalid input."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INTERNAL_ERROR or CONFIGURATION_ERROR. Static, sanitized detail."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "JOB_STORE_UNAVAILABLE, KNOWLEDGE_UNAVAILABLE, APPROVAL_STORE_UNAVAILABLE or HISTORY_STORE_UNAVAILABLE. Static, sanitized detail."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Evaluation Status",
        "tags": [
          "evaluations"
        ],
        "x-ownership": "Owner only for analyst/reviewer; admin may access all owned jobs.",
        "x-roles": [
          "analyst",
          "reviewer",
          "admin"
        ]
      }
    },
    "/api/v1/evaluations/{job_id}/cancel": {
      "post": {
        "description": "Roles: admin. Owner only for analyst/reviewer; admin may access all owned jobs. Identity and current role come from the validated JWT and stored user.",
        "operationId": "cancel_evaluation_api_v1_evaluations__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "DOMAIN_ERROR, APPLICATION_ERROR or HTTP_ERROR: invalid operation."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "NOT_AUTHENTICATED. WWW-Authenticate: Bearer.",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "const": "Bearer",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "PERMISSION_DENIED or RESOURCE_FORBIDDEN."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RESOURCE_NOT_FOUND or HTTP_ERROR."
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "HTTP_ERROR: method not allowed."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INVALID_STATE_TRANSITION: conflict with persisted state."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "UPLOAD_TOO_LARGE."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "VALIDATION_ERROR or INVARIANT_VIOLATION: invalid input."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INTERNAL_ERROR or CONFIGURATION_ERROR. Static, sanitized detail."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "JOB_STORE_UNAVAILABLE, KNOWLEDGE_UNAVAILABLE, APPROVAL_STORE_UNAVAILABLE or HISTORY_STORE_UNAVAILABLE. Static, sanitized detail."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Cancel Evaluation",
        "tags": [
          "evaluations"
        ],
        "x-ownership": "Owner only for analyst/reviewer; admin may access all owned jobs.",
        "x-roles": [
          "admin"
        ]
      }
    },
    "/api/v1/evaluations/{job_id}/report": {
      "get": {
        "description": "Roles: analyst, reviewer, admin. Owner only for analyst/reviewer; admin may access all owned jobs. Identity and current role come from the validated JWT and stored user.",
        "operationId": "evaluation_report_api_v1_evaluations__job_id__report_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "json",
              "enum": [
                "json",
                "markdown"
              ],
              "title": "Format",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "DOMAIN_ERROR, APPLICATION_ERROR or HTTP_ERROR: invalid operation."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "NOT_AUTHENTICATED. WWW-Authenticate: Bearer.",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "const": "Bearer",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "PERMISSION_DENIED or RESOURCE_FORBIDDEN."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RESOURCE_NOT_FOUND or HTTP_ERROR."
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "HTTP_ERROR: method not allowed."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INVALID_STATE_TRANSITION: conflict with persisted state."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "UPLOAD_TOO_LARGE."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "VALIDATION_ERROR or INVARIANT_VIOLATION: invalid input."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INTERNAL_ERROR or CONFIGURATION_ERROR. Static, sanitized detail."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "JOB_STORE_UNAVAILABLE, KNOWLEDGE_UNAVAILABLE, APPROVAL_STORE_UNAVAILABLE or HISTORY_STORE_UNAVAILABLE. Static, sanitized detail."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Evaluation Report",
        "tags": [
          "evaluations"
        ],
        "x-ownership": "Owner only for analyst/reviewer; admin may access all owned jobs.",
        "x-roles": [
          "analyst",
          "reviewer",
          "admin"
        ]
      }
    },
    "/api/v1/evaluations/{job_id}/restart": {
      "post": {
        "description": "Roles: admin. Owner only for analyst/reviewer; admin may access all owned jobs. Identity and current role come from the validated JWT and stored user.",
        "operationId": "restart_evaluation_api_v1_evaluations__job_id__restart_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "DOMAIN_ERROR, APPLICATION_ERROR or HTTP_ERROR: invalid operation."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "NOT_AUTHENTICATED. WWW-Authenticate: Bearer.",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "const": "Bearer",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "PERMISSION_DENIED or RESOURCE_FORBIDDEN."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RESOURCE_NOT_FOUND or HTTP_ERROR."
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "HTTP_ERROR: method not allowed."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INVALID_STATE_TRANSITION: conflict with persisted state."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "UPLOAD_TOO_LARGE."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "VALIDATION_ERROR or INVARIANT_VIOLATION: invalid input."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INTERNAL_ERROR or CONFIGURATION_ERROR. Static, sanitized detail."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "JOB_STORE_UNAVAILABLE, KNOWLEDGE_UNAVAILABLE, APPROVAL_STORE_UNAVAILABLE or HISTORY_STORE_UNAVAILABLE. Static, sanitized detail."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Restart Evaluation",
        "tags": [
          "evaluations"
        ],
        "x-ownership": "Owner only for analyst/reviewer; admin may access all owned jobs.",
        "x-roles": [
          "admin"
        ]
      }
    },
...
  }
}